+++ packages/central-server/src/permissions/assertions.js
@@ -12,8 +12,29 @@ export {
   assertVizBuilderAccess,
   hasTupaiaAdminPanelAccess,
   hasTupaiaAdminPanelAccessToCountry,
-  assertAdminPanelAccessToCountry,
   assertAdminPanelAccess,
   assertPermissionGroupAccess,
   assertPermissionGroupsAccess,
 } from '@tupaia/access-policy';
+
+/**
+ * @type {PermissionsAssertion}
+ * @param {ModelRegistry} models
+ * @param {string} recordId
+ */
+export const assertAdminPanelAccessToCountry = async (accessPolicy, models, recordId) => {
+  const entity = ensure(
+    await models.entity.findById(recordId),
+    `No entity exists with ID ${recordId}`,
+  );
+  const userHasAdminAccessToCountry = accessPolicy.allows(
+    entity.country_code,
+    TUPAIA_ADMIN_PANEL_PERMISSION_GROUP,
+  );
+  if (!userHasAdminAccessToCountry) {
+    throw new PermissionsError(
+      `Need ${TUPAIA_ADMIN_PANEL_PERMISSION_GROUP} access to country ‘${entity.country_code}’ to edit entity ‘${entity.name}’`,
+    );
+  }
+  return true;
+};
